% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.coef.R
\name{beta.coef}
\alias{beta.coef}
\alias{beta.coef.default}
\title{Beta coefficients}
\usage{
beta.coef(object, ...)

\method{beta.coef}{default}(object, complete = FALSE, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
meaningful.}

\item{...}{further arguments passed to or from other methods.}

\item{complete}{for the default (used for lm, etc) and aov methods: logical
indicating if the full coefficient vector should be returned also in case of
an over-determined system where some coefficients will be set to \link{NA}.}
}
\description{
Computes the beta (regression) coefficients, also called standardized
coefficients or beta weights, to quantify the importance (the effect) of the
predictors on the dependent variable in a multiple regression analysis where
the variables are measured in different units.
}
\details{
The beta weights are the coefficient estimates resulting from a regression
analysis where the underlying data have been standardized so that the
variances of dependent and explanatory variables are equal to 1.
Therefore, standardized coefficients are unitless and refer to how many
standard deviations a dependent variable will change, per standard deviation
increase in the predictor variable.
See \url{https://en.wikipedia.org/wiki/Standardized_coefficient} or
\href{https://rdrr.io/cran/QuantPsyc/man/lm.beta.html}{QuantPsyc::lm.beta}.

Based on \code{QuantPsyc::lm.beta()}
}
\examples{
fit <- lm(fidelida ~ velocida + calidadp, hbat)
coef(fit)
beta.coef(fit)
fit2 <- lm(scale(fidelida) ~ scale(velocida) + scale(calidadp), hbat)
coef(fit2)
}
\seealso{
\code{\link[=coef]{coef()}}
}
