% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{obs.pred.plot}
\alias{obs.pred.plot}
\title{Observed vs. predicted plot}
\usage{
obs.pred.plot(
  pred,
  obs,
  xlab = "Predicted",
  ylab = "Observed",
  lm.fit = TRUE,
  lowess = TRUE,
  ...
)
}
\arguments{
\item{pred}{a numeric vector with the predicted values.}

\item{obs}{a numeric vector with the observed values.}

\item{xlab}{a title for the axis corresponding to the predicted values.}

\item{ylab}{a title for the axis corresponding to the observed values.}

\item{lm.fit}{logical indicating if a \code{\link[stats]{lm}} fit is
added to the plot.}

\item{lowess}{logical indicating if a \code{\link[stats]{lowess}} smooth is
added to the plot.}

\item{...}{additional graphical parameters.}
}
\value{
Invisibly returns the fitted linear model if \code{lm.fit == TRUE}.
}
\description{
Draws a scatter plot of the observed values against the predicted values.
}
\examples{
set.seed(1)
nobs <- nrow(hbat)
itrain <- sample(nobs, 0.8 * nobs)
train <- hbat[itrain, ]
test <- hbat[-itrain, ]
fit <- lm(fidelida ~ velocida + calidadp, train)
pred <- predict(fit, newdata = test)
obs <- test$fidelida
res <- obs.pred.plot(pred, obs)
summary(res)
}
